requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.5
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
zope.interface==5.4.0
